Add files via upload
This java file takes an input as a csv file from the accounting website/app: intuit quickbooks. This specific code takes a csv of the profit and loss data of a company (pnl data) and consolidates the data to a printed report that is easy to read. It also checks the data and makes sure that the totals are correct in all categories of the report. The output of the program is a text document.